Convert a Data Frame to a Named List of Vectors

Convert a Data Frame to a Named List of Vectors

Usage

df2list(df, name_col, value_col, clean_names = FALSE)

Arguments

df

Data frame to be converted.

name_col

Character. Name of the column in df to be used for list names.

value_col

Character. Name of the column in df to be used as list elements (vectors).

clean_names

Logical. Should list names be modified to remove whitespaces and non-aphanumeric strings (excluding underscores)? See SomaDataIO::cleanNames() for more details.

Value

A list of vectors.

Author

Amanda Hiser

Examples

df <- SomaDataIO::rn2col(SomaDataIO::example_data, "Rowname")
df2list(df, name_col = "SampleType", value_col = "Rowname")
#> $Sample
#>   [1] "258495800012_3" "258495800004_7" "258495800010_8"
#>   [4] "258495800003_4" "258495800009_4" "258495800012_8"
#>   [7] "258495800001_3" "258495800004_8" "258495800001_8"
#>  [10] "258495800009_8" "258495800012_5" "258495800006_2"
#>  [13] "258495800012_1" "258495800007_3" "258495800011_3"
#>  [16] "258495800002_2" "258495800005_8" "258495800003_2"
#>  [19] "258495800005_2" "258495800008_4" "258495800010_2"
#>  [22] "258495800010_5" "258495800006_6" "258495800012_4"
#>  [25] "258495800001_7" "258495800007_8" "258495800011_7"
#>  [28] "258495800002_6" "258495800006_8" "258495800003_6"
#>  [31] "258495800005_5" "258495800008_7" "258495800011_5"
#>  [34] "258495800005_6" "258495800012_7" "258495800001_2"
#>  [37] "258495800008_8" "258495800011_1" "258495800005_7"
#>  [40] "258495800002_5" "258495800006_5" "258495800011_6"
#>  [43] "258495800009_3" "258495800007_1" "258495800006_3"
#>  [46] "258495800012_2" "258495800009_7" "258495800004_2"
#>  [49] "258495800001_4" "258495800004_4" "258495800001_1"
#>  [52] "258495800012_6" "258495800009_1" "258495800007_7"
#>  [55] "258495800003_5" "258495800009_5" "258495800004_6"
#>  [58] "258495800010_6" "258495800008_6" "258495800005_4"
#>  [61] "258495800003_7" "258495800002_7" "258495800011_8"
#>  [64] "258495800008_1" "258495800007_4" "258495800003_3"
#>  [67] "258495800009_2" "258495800004_5" "258495800010_3"
#>  [70] "258495800005_3" "258495800003_1" "258495800006_1"
#>  [73] "258495800002_1" "258495800011_2" "258495800007_2"
#>  [76] "258495800008_3" "258495800002_4" "258495800004_1"
#>  [79] "258495800011_4" "258495800007_5" "258495800006_4"
#>  [82] "258495800005_1" "258495800002_8" "258495800010_1"
#>  [85] "258495800008_2" "258495800104_3" "258495800112_7"
#>  [88] "258495800103_7" "258495800101_6" "258495800112_2"
#>  [91] "258495800109_4" "258495800111_8" "258495800102_5"
#>  [94] "258495800111_4" "258495800101_4" "258495800108_8"
#>  [97] "258495800105_3" "258495800106_3" "258495800110_8"
#> [100] "258495800104_7" "258495800103_1" "258495800107_4"
#> [103] "258495800110_2" "258495800101_1" "258495800112_5"
#> [106] "258495800101_8" "258495800108_1" "258495800105_8"
#> [109] "258495800106_7" "258495800111_5" "258495800104_8"
#> [112] "258495800103_2" "258495800110_3" "258495800107_8"
#> [115] "258495800109_7" "258495800101_3" "258495800102_6"
#> [118] "258495800107_3" "258495800105_2" "258495800108_2"
#> [121] "258495800110_4" "258495800105_6" "258495800102_3"
#> [124] "258495800111_3" "258495800106_6" "258495800101_7"
#> [127] "258495800112_3" "258495800106_4" "258495800101_5"
#> [130] "258495800112_1" "258495800109_1" "258495800101_2"
#> [133] "258495800104_1" "258495800108_4" "258495800108_7"
#> [136] "258495800107_6" "258495800103_5" "258495800112_8"
#> [139] "258495800103_3" "258495800110_1" "258495800102_4"
#> [142] "258495800105_7" "258495800107_1" "258495800111_6"
#> [145] "258495800106_2" "258495800103_8" "258495800109_3"
#> [148] "258495800105_1" "258495800110_6" "258495800102_8"
#> [151] "258495800109_8" "258495800107_5" "258495800102_1"
#> [154] "258495800106_1" "258495800110_7" "258495800106_8"
#> [157] "258495800103_6" "258495800109_2" "258495800104_5"
#> [160] "258495800111_2" "258495800106_5" "258495800102_7"
#> [163] "258495800104_4" "258495800107_2" "258495800109_5"
#> [166] "258495800108_3" "258495800104_2" "258495800108_5"
#> [169] "258495800105_5" "258495800110_5"
#> 
#> $Calibrator
#>  [1] "258495800004_3" "258495800010_7" "258495800001_5"
#>  [4] "258495800006_7" "258495800008_5" "258495800108_6"
#>  [7] "258495800111_1" "258495800104_6" "258495800111_7"
#> [10] "258495800102_2"
#> 
#> $Buffer
#> [1] "258495800010_4" "258495800003_8" "258495800009_6" "258495800103_4"
#> [5] "258495800112_4" "258495800105_4"
#> 
#> $QC
#> [1] "258495800001_6" "258495800007_6" "258495800002_3" "258495800109_6"
#> [5] "258495800112_6" "258495800107_7"
#> 

# Use clean_names argument to make R-compatible list names
df <- SomaDataIO::getAnalyteInfo(SomaDataIO::example_data)
head(df$TargetFullName)
#> [1] "Beta-crystallin B2"                                
#> [2] "RAF proto-oncogene serine/threonine-protein kinase"
#> [3] "Zinc finger protein 41"                            
#> [4] "ETS domain-containing protein Elk-1"               
#> [5] "Guanylyl cyclase-activating protein 1"             
#> [6] "Inositol polyphosphate 5-phosphatase OCRL-1"       
df2list(df, name_col = "TargetFullName", value_col = "AptName",
        clean_names = TRUE) |>
head()
#> $Beta.crystallin.B2
#> [1] "seq.10000.28"
#> 
#> $RAF.proto.oncogene.serine.threonine.protein.kinase
#> [1] "seq.10001.7"
#> 
#> $Zinc.finger.protein.41
#> [1] "seq.10003.15"
#> 
#> $ETS.domain.containing.protein.Elk.1
#> [1] "seq.10006.25"
#> 
#> $Guanylyl.cyclase.activating.protein.1
#> [1] "seq.10008.43"
#> 
#> $Inositol.polyphosphate.5.phosphatase.OCRL.1
#> [1] "seq.10011.65"
#>